+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
   - Button for hosts to add the quiz they are on to the queue.
   - Draggable (for hosts) list of quizzes in the queue.
   - Remove button for hosts to remove quizzes from the queue.
+  - Go to quiz button for hosts to visit the page of quizzes in the queue.
 
 [v2.2.0] - 2021-08-16
 ---------------------